Break item property classes into a separate file
