+++ .pre-commit-config.yaml
@@ -1,17 +1,16 @@
 repos:
-- repo: https://github.com/pre-commit/pre-commit-hooks 
-  rev: v3.2.0 
-  hooks: 
-  - id: trailing-whitespace 
-  - id: end-of-file-fixer 
+- repo: https://github.com/pre-commit/pre-commit-hooks
+  rev: v3.2.0
+  hooks:
+  - id: trailing-whitespace
+  - id: end-of-file-fixer
   - id: debug-statements
-- repo: https://github.com/psf/black 
-  rev: 23.3.0 
-  hooks: 
-  - id: black 
-    language_version: python3.11 
-- repo: https://github.com/pycqa/isort 
-  rev: 5.12.0 
-  hooks: 
-  - id: isort 
-  
+- repo: https://github.com/psf/black
+  rev: 23.3.0
+  hooks:
+  - id: black
+    language_version: python3.11
+- repo: https://github.com/pycqa/isort
+  rev: 5.12.0
+  hooks:
+  - id: isort
