+++ pyproject.toml
@@ -1,8 +1,8 @@
 [tool.poetry]
-name = "acorectl"
+name = "hogger"
 version = "0.1.0"
 description = ""
-authors = ["Your Name <you@example.com>"]
+authors = ["Nick Saccente"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
