+++ src/items/__init__.py
@@ -1,1 +1,3 @@
-from .weapon import OneHandedAxe, TwoHandedAxe
+from src.items.item import Item
+from src.items.item_properties import *
+from src.items.weapon import OneHandedAxe, TwoHandedAxe
