+++ src/items/item.py
@@ -1,231 +1,10 @@
 import abc
-from enum import Enum
 
 from pydantic import BaseModel, Field, FieldValidationInfo, field_validator
 
-from src.misc import Money
+from src.misc import Money, Time
 
-
-class Quality(Enum):
-    Poor = 0
-    Common = 1
-    Uncommon = 2
-    Rare = 3
-    Epic = 4
-    Legendary = 5
-    Artifact = 6
-    BoA = 7
-
-
-class ItemClass(Enum):
-    Consumable = 0
-    Container = 1
-    Weapon = 2
-    Gem = 3
-    Armor = 4
-    Reagent = 5
-    Projectile = 6
-    TradeGoods = 7
-    Generic_OBSELETE = 8
-    Money_OBSELETE = 9
-    Quiver = 11
-    Quest = 12
-    Key = 13
-    Permanent_OBSELETE = 14
-    Misc = 15
-    Glyph = 16
-
-
-class InventoryType(Enum):
-    NoEquip = 0
-    Head = 1
-    Neck = 2
-    Shoulders = 3
-    Body = 4
-    Chest = 5
-    Waist = 6
-    Legs = 7
-    Feet = 8
-    Wrists = 9
-    Hands = 10
-    Finger = 11
-    Trinket = 12
-    WeaponOneHanded = 13
-    Shield = 14
-    Ranged = 15
-    Cloak = 16
-    WeaponTwoHanded = 17
-    Bag = 18
-    Tabard = 19
-    Robe = 20
-    MainHand = 21
-    OffHand = 22
-    Holdable = 23
-    Ammo = 24
-    Thrown = 25
-    RangedRight = 26
-    Quiver = 27
-    Relic = 28
-
-
-class Material(Enum):
-    Consumables = -1
-    Undefined = 0
-    Metal = 1
-    Wood = 2
-    Liquid = 3
-    Jewelry = 4
-    Chain = 5
-    Plate = 6
-    Cloth = 7
-    Leather = 8
-
-
-class ItemBinding(Enum):
-    Never = 0
-    OnPickup = 1
-    OnEquip = 2
-    OnUse = 3
-    QuestItem = 4
-    QuestItem1 = 5
-
-
-class FoodType(Enum):
-    Undefined = 0
-    Meat = 1
-    Fish = 2
-    Cheese = 3
-    Bread = 4
-    Fungus = 5
-    Fruit = 6
-    RawMeat = 7
-    RawFish = 8
-
-
-class TotemCategory(Enum):
-    Undefined = 0
-    SkinningKnife_OLD = 1
-    EarthTotem = 2
-    AirTotem = 3
-    FireTotem = 4
-    WaterTotem = 5
-    RunedCopperRod = 6
-    RunedSilverRod = 7
-    RunedGoldenod = 8
-    RunedTruesilverRod = 9
-    RunedArcaniteRod = 10
-    MiningPick_OLD = 11
-    PhilosophersStone = 12
-    BlacksmithHammer_OLD = 13
-    ArclightSpanner = 14
-    GyromaticMicroAdjustor = 15
-    MasterTotem = 21
-    RunedFelIronRod = 41
-    RunedAdamantiteRod = 62
-    RunedEterniumRod = 63
-    HollowQuill = 81
-    RunedAzuriteRod = 101
-    VirtuosoInkingSet = 121
-    Drums = 141
-    GnomishArmyKnife = 161
-    BlacksmithHammer = 162
-    MiningPick = 165
-    SkinningKnife = 166
-    HammerPick = 167
-    BladedPickaxe = 168
-    FlintAndTinder = 169
-    RunedCobaltRod = 189
-    RunedTitaniumRod = 190
-
-
-class ItemStat(Enum):
-    Mana = 0
-    Health = 1
-    Agility = 3
-    Strength = 4
-    Intellect = 5
-    Spirit = 6
-    Stamina = 7
-    DefenseRating = 12
-    DodgeRating = 13
-    ParryRating = 14
-    BlockRating = 15
-    MeleeHitRating = 16
-    RangedHitRating = 17
-    SpellHitRating = 18
-    MeleeCritRating = 19
-    RangedCritRating = 20
-    SpellCritRating = 21
-    MeleeAvoidanceRating = 22
-    RangedAvoidanceRating = 23
-    SpellAvoidanceRating = 24
-    MeleeCritAvoidanceRating = 25
-    RangedCritAvoidanceRating = 26
-    SpellCritAvoidanceRating = 27
-    MeleeHasteRating = 28
-    RangedHasteRating = 29
-    SpellHasteRating = 30
-    HitRating = 31
-    CritRating = 32
-    HitAvoidanceRating = 33
-    CritAvoidanceRating = 34
-    ResilienceRating = 35
-    HasteRating = 36
-    ExpertiseRating = 37
-    AttackPower = 38
-    RangedAttackPower = 39
-    FeralAttackPower_OLD = 40
-    SpellHealing = 41
-    SpellDamage = 42
-    ManaRegen = 43
-    ArmorPenetration = 44
-    SpellPower = 45
-    HealthRegen = 46
-    SpellPenetration = 47
-    Block = 48
-
-
-class BagFamily(BaseModel):
-    Arrows: bool = False
-    Bullets: bool = False
-    SoulShards: bool = False
-    Leatherworking: bool = False
-    Inscription: bool = False
-    Herbs: bool = False
-    Enchanting: bool = False
-    Engineering: bool = False
-    Keys: bool = False
-    Gems: bool = False
-    Mining: bool = False
-    SoulboundEquipment: bool = False
-    VanityPets: bool = False
-    CurrencyTokens: bool = False
-    QuestItems: bool = False
-
-    def __setitem__(self, key, value):
-        setattr(self, key, value)
-
-    def __getitem__(self, key):
-        return getattr(self, key)
-
-    def __int__(self):
-        return (
-            (self.Arrows * 1)
-            + (self.Bullets * 2)
-            + (self.SoulShards * 4)
-            + (self.Leatherworking * 8)
-            + (self.Inscription * 16)
-            + (self.Herbs * 32)
-            + (self.Enchanting * 64)
-            + (self.Engineering * 128)
-            + (self.Keys * 256)
-            + (self.Gems * 512)
-            + (self.Mining * 1024)
-            + (self.SoulboundEquipment * 2048)
-            + (self.VanityPets * 4096)
-            + (self.CurrencyTokens * 8192)
-            + (self.QuestItems * 16384)
-        )
+from .item_properties import *
 
 
 class Item(BaseModel, abc.ABC):
@@ -250,7 +29,7 @@ class Item(BaseModel, abc.ABC):
             "overriden by the unique sound of a different weapon type"
         ),
     )
-    displayID: int = Field(
+    displayId: int = Field(
         description="Controls both the model appearance and icon.",
     )
     quality: Quality = Field(
@@ -282,7 +61,7 @@ class Item(BaseModel, abc.ABC):
         description="Is the item equippable? A quest item?",
     )
     maxAmount: int = Field(
-        default=0,
+        default=1,
         description="The maximum amount that a player can have; use 0 for infinite",
     )
     stackSize: int = Field(
@@ -304,12 +83,15 @@ class Item(BaseModel, abc.ABC):
         default=BagFamily(),
         description="Dictates what kind of bags this item can be placed in.",
     )
+    # randomItemProperty
     # randomPropertyItem with RandomProperty = 863
     # Item with RandomSuffix = 24585
     # min(RandomProperty, RandomSuffix) must equal 0.
     containerSlots: int = Field(
         default=0,
-        description="If this item is a bag, controls the number of slots it will have",
+        description=(
+            "If this item is a bag, controls the number of slots it will have"
+        ),
     )
     specialUse: TotemCategory = Field(
         default=TotemCategory.Undefined,
@@ -318,6 +100,26 @@ class Item(BaseModel, abc.ABC):
             "shaman's totems, "
         ),
     )
+    duration: Time = Field(
+        default=Time(),
+        description=(
+            "The amount of time an item will exist in a player's inventory "
+            "before disappearing; setting the duration to 0 seconds will "
+            "prevent the item from every disappearing"
+        ),
+    )
+    itemLimitCategory: int = Field(
+        default=0,
+        description=(
+            'defines if an item belongs to a "category", like "Mana Gems" or '
+            'Healthstone" and it defines how many items of the category you '
+            'can have in the bag (this is the "limit"). For example, for '
+            'Healthstone, there are several items like "Lesser Healthstone, '
+            'Greater Healthstone, etc." but you can have only one in your bag '
+            "(check as example value 3 or 4)."
+        ),
+    )
+    # disenchantId = <SELECT * FROM `item_template` WHERE `itemLimitCategory`>
 
     @field_validator("buyCount", "stackSize")
     def integer_greater_than_zero(cls, v: int, info: FieldValidationInfo) -> int:
