+++ src/misc/__init__.py
@@ -1,1 +1,1 @@
-from .money import Money
+from .currency import Money, Time
